+++ Gateway/Validator/Gateway/In3Validator.php
@@ -17,66 +17,12 @@ declare(strict_types=1);
 
 namespace MultiSafepay\ConnectCore\Gateway\Validator\Gateway;
 
-use Magento\Payment\Gateway\Validator\AbstractValidator;
 use Magento\Payment\Gateway\Validator\ResultInterface;
-use Magento\Payment\Gateway\Validator\ResultInterfaceFactory;
-use MultiSafepay\ConnectCore\Model\Api\Validator\DateOfBirthValidator;
-use MultiSafepay\ConnectCore\Model\Api\Validator\GenderValidator;
 
-class In3Validator extends AbstractValidator
+class In3Validator extends BaseGatewayValidator
 {
-    /**
-     * @var DateOfBirthValidator
-     */
-    private $dateOfBirthValidator;
-
-    /**
-     * @var GenderValidator
-     */
-    private $genderValidator;
-
-    /**
-     * In3Validator constructor.
-     *
-     * @param DateOfBirthValidator $dateOfBirthValidator
-     * @param GenderValidator $genderValidator
-     * @param ResultInterfaceFactory $resultFactory
-     */
-    public function __construct(
-        DateOfBirthValidator $dateOfBirthValidator,
-        GenderValidator $genderValidator,
-        ResultInterfaceFactory $resultFactory
-    ) {
-        $this->dateOfBirthValidator = $dateOfBirthValidator;
-        $this->genderValidator = $genderValidator;
-        parent::__construct($resultFactory);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function validate(array $validationSubject): ResultInterface
-    {
-        $payment = $validationSubject['payment'] ?? null;
-
-        if (!$payment) {
-            return $this->createResult(false, [__('Can\'t get a payment information')]);
-        }
-
-        if ($paymentAdditionalInformation = $payment->getAdditionalInformation()) {
-            if (!isset($paymentAdditionalInformation['date_of_birth'])
-                || !$this->dateOfBirthValidator->validate($paymentAdditionalInformation['date_of_birth'])
-            ) {
-                return $this->createResult(false, [__('Invalid Date of Birth')]);
-            }
-
-            if (!isset($paymentAdditionalInformation['gender'])
-                || !$this->genderValidator->validate($paymentAdditionalInformation['gender'])
-            ) {
-                return $this->createResult(false, [__('Please choose a gender')]);
-            }
-        }
-
-        return $this->createResult(true);
-    }
+    public const AVAILABLE_VALIDATORS = [
+        'date_of_birth',
+        'gender'
+    ];
 }